requirements.txt
pandas
openpyxl
streamlit
pyyaml
pandera
loguru
xlsxwriter
pytest
plotly
